+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 12.3
 ---------------
 - Multilingual course title: use ParseMLField(), program wide
 - Multilingual course title: SQL change title column type from varchar(100) to text in Update.fnc.php, rosariosis.sql & rosariosis_mysql.sql
+- Multilingual course title: use MLTextInput() in Courses.php
 
 Changes in 12.2.3
 -----------------